backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.23
langgraph==0.0.47
langsmith>=0.0.77,<0.1.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
sentencepiece
numpy>=1.24.0